requirements.txt
# Backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-groq>=0.1.0
faiss-cpu>=1.7.4
pypdf>=3.17.0
sentence-transformers>=2.2.2
torch>=2.0.0
python-dotenv>=1.0.0

# Optional: Keep streamlit for legacy support
streamlit>=1.28.0

